rename ks8util_test to test and add documentation
