+++ pkg/util/k8sutil/errors_test.go
@@ -1,3 +1,25 @@
+//
+// DISCLAIMER
+//
+// Copyright 2018 ArangoDB GmbH, Cologne, Germany
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Copyright holder is ArangoDB GmbH, Cologne, Germany
+//
+// Author Jan Christoph Uhde
+//
+
 package k8sutil
 
 import (
@@ -20,7 +42,7 @@ var (
 
 func TestIsAlreadyExists(t *testing.T) {
 	assert.False(t, IsAlreadyExists(conflictError))
-	assert.False(t, IsConflict(maskAny(invalidError)))
+	assert.False(t, IsAlreadyExists(maskAny(invalidError)))
 	assert.True(t, IsAlreadyExists(existsError))
 	assert.True(t, IsAlreadyExists(maskAny(existsError)))
 }
@@ -34,7 +56,7 @@ func TestIsConflict(t *testing.T) {
 
 func TestIsNotFound(t *testing.T) {
 	assert.False(t, IsNotFound(invalidError))
-	assert.False(t, IsConflict(maskAny(invalidError)))
+	assert.False(t, IsNotFound(maskAny(invalidError)))
 	assert.True(t, IsNotFound(notFoundError))
 	assert.True(t, IsNotFound(maskAny(notFoundError)))
 }
