+++ pkg/util/k8sutil/test/events_test.go
@@ -1,4 +1,26 @@
-package k8sutil_test
+//
+// DISCLAIMER
+//
+// Copyright 2018 ArangoDB GmbH, Cologne, Germany
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Copyright holder is ArangoDB GmbH, Cologne, Germany
+//
+// Author Jan Christoph Uhde
+//
+
+package test
 
 import (
 	"errors"
@@ -56,9 +78,3 @@ func TestErrorEvent(t *testing.T) {
 	assert.Equal(t, event.Reason, "Reason")
 	assert.Equal(t, event.Type, v1.EventTypeWarning)
 }
-
-// // not accessible outside the package
-// func TestDeploymentEvent(t *testing.T) {
-// 	event := k8sutil.New("member name", "role", &apiObjectForTest)
-// 	assert.Equal(t, event.Type, v1.EventTypeNormal)
-// }
